requirements.txt
Django==3.2.8
django_ckeditor==5.9.0
django_rest_swagger==2.2.0
django_simple_history==3.0.0
djangorestframework==3.13.1
djangorestframework_simplejwt==5.0.0
stripe==2.64.0
typing_extensions==3.10.0.2
django-cors-headers==3.10.1
Pillow==8.4.0
pip==21.1.2
PyJWT
pytz==2021.3
requests==2.26.0
setuptools==57.0.0
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
wheel==0.36.2
gunicorn==20.1.0
whitenoise==5.3.0